Move openPopup to separated modular JS function and call it on import in renderCards
